backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "civet"
version = "0.1.0"
description = "A FastAPI application with SQLAlchemy and Alembic"
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "fastapi[standard]>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "openai==1.70.0",
    "google-cloud-vision==3.10.1",
    "asyncpg==0.30.0",
    "greenlet==3.1.1",
    "psycopg2==2.9.10",
    "python-dateutil==2.9.0.post0",
    "minio==7.2.15",
    "pytz==2025.2",
    "python-jose==3.4.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0", "ruff>=0.3.0"]

[tool.setuptools]
packages = ["app"]

[tool.hatch.build.targets.wheel]
packages = ["app/main"]
